Don’t check in test_db.sqlite3
