+++ CreeDictionary/.gitignore
@@ -1,0 +1,1 @@
+/test_db.sqlite3
