+++ CreeDictionary/tests/DatabaseManager_tests/xml_importer_test.py
@@ -2,7 +2,7 @@ import pytest
 from API.models import Wordform
 from DatabaseManager.cree_inflection_generator import expand_inflections
 from DatabaseManager.xml_importer import find_latest_xml_file
-from tests.conftest import migrate_and_import
+from tests.DatabaseManager_tests.conftest import migrate_and_import
 from utils import PartOfSpeech, shared_res_dir
